BCFile::getMethodProperties(): sync with PHPCS / union type support
PR 168 added support for union types to the `FunctionDeclarations::getProperties()` method and by extension the `BCFile::getMethodProperties()` method, even though PHPCS upstream did not support union types yet.

Upstream PR 3032 added support for union types to the PHPCS native `File::getMethodProperties()` method.

As that PR has now been merged, this commit syncs the upstream changes in, by moving the unit tests from the "Diff" tests to the BCFile tests and annotating that union types are now officially supported in the `BCFile::getMethodProperties()` method.
